\name{ join }

\alias{join}

\title{ Join data }

\description{
  Join abbreviations form a taxonomic reference list.
}

\usage{
join(x, y, make.names = TRUE)
}

\arguments{
  \item{ x }{
    \code{data.frame}.
  }
  \item{ y }{
    \code{data.frame}.}
  \item{ make.names }{
    \code{logical}.}

}

\details{
  Nothing here yet!
}

\value{
data.frame
}

\author{ Roland Kaiser }

\examples{
# Nothing here yet
}

\keyword{ manip }
