\name{vegit-package}
\alias{vegit-package}
\alias{vegit}
\docType{package}

\title{ Maintain, extend and mange the vegsoup-data repository }

\description{
Functions to maintain, extend and create data inside and for the (\href{https://github.com/kardinal-eros/vegsoup-data}{vegsoup-data}) repository.
}

\details{
\tabular{ll}{
Package: \tab voucher\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2014-03-02\cr
License: \tab GPL (>= 2)\cr
}

A bunch of utility functions.
}

\author{
Maintainer:  Roland Kaiser <kardinal.eros@gemail.com>
}

%\references{
%}

\keyword{ package }
\seealso{
\code{\link[vegsoup:vegsoup-package]{vegsoup}}
}
%\examples{
%}
